[Modify] 상품 목록 페이지에서 렌더링되는 productItem 컴포넌트 분리작업 - productItem을 ProductItemImage, ProductItemInfo로 분리
#23 완료
